+++ src/globus_cli/commands/bookmark/create.py
@@ -2,8 +2,8 @@ import click
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import ENDPOINT_PLUS_REQPATH, command
-from globus_cli.safeio import formatted_print
 from globus_cli.services.transfer import TRANSFER_RESOURCE_SERVER, get_client
+from globus_cli.termio import formatted_print
 
 
 @command(
