+++ src/globus_cli/commands/delete.py
@@ -10,16 +10,16 @@ from globus_cli.parsing import (
     shlex_process_stdin,
     task_submission_options,
 )
-from globus_cli.safeio import (
-    FORMAT_TEXT_RECORD,
-    err_is_terminal,
-    formatted_print,
-    term_is_interactive,
-)
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     autoactivate,
     get_client,
+)
+from globus_cli.termio import (
+    FORMAT_TEXT_RECORD,
+    err_is_terminal,
+    formatted_print,
+    term_is_interactive,
 )
