+++ src/globus_cli/commands/endpoint/activate.py
@@ -8,12 +8,12 @@ from globus_cli.helpers import (
 )
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command, endpoint_id_arg, mutex_option_group
-from globus_cli.safeio import FORMAT_TEXT_RAW, formatted_print
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     activation_requirements_help_text,
     get_client,
 )
+from globus_cli.termio import FORMAT_TEXT_RAW, formatted_print
 
 
 @command(
