+++ src/globus_cli/commands/endpoint/is_activated.py
@@ -2,12 +2,12 @@ import click
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command, endpoint_id_arg
-from globus_cli.safeio import formatted_print
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     activation_requirements_help_text,
     get_client,
 )
+from globus_cli.termio import formatted_print
 
 
 @command(
