+++ src/globus_cli/commands/endpoint/my_shared_endpoint_list.py
@@ -1,11 +1,11 @@
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command, endpoint_id_arg
-from globus_cli.safeio import formatted_print
 from globus_cli.services.transfer import (
     ENDPOINT_LIST_FIELDS,
     TRANSFER_RESOURCE_SERVER,
     get_client,
 )
+from globus_cli.termio import formatted_print
 
 
 @command(
