+++ src/globus_cli/commands/endpoint/permission/create.py
@@ -2,13 +2,13 @@ import click
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import ENDPOINT_PLUS_REQPATH, command, security_principal_opts
-from globus_cli.safeio import FORMAT_TEXT_RECORD, formatted_print
 from globus_cli.services.auth import AUTH_RESOURCE_SERVER, maybe_lookup_identity_id
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     assemble_generic_doc,
     get_client,
 )
+from globus_cli.termio import FORMAT_TEXT_RECORD, formatted_print
 
 
 @command(
