+++ src/globus_cli/commands/endpoint/permission/list.py
@@ -2,9 +2,9 @@ from globus_sdk import IdentityMap
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command, endpoint_id_arg
-from globus_cli.safeio import formatted_print
 from globus_cli.services.auth import AUTH_RESOURCE_SERVER, get_auth_client
 from globus_cli.services.transfer import TRANSFER_RESOURCE_SERVER, get_client
+from globus_cli.termio import formatted_print
 
 
 @command(
