+++ src/globus_cli/commands/endpoint/server/show.py
@@ -2,8 +2,8 @@ from textwrap import dedent
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command, endpoint_id_arg, server_id_arg
-from globus_cli.safeio import FORMAT_TEXT_RECORD, formatted_print
 from globus_cli.services.transfer import TRANSFER_RESOURCE_SERVER, get_client
+from globus_cli.termio import FORMAT_TEXT_RECORD, formatted_print
 
 
 @command(
