+++ src/globus_cli/commands/endpoint/set_subscription_id.py
@@ -4,8 +4,8 @@ import click
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import EXPLICIT_NULL, command, endpoint_id_arg
-from globus_cli.safeio import FORMAT_TEXT_RAW, formatted_print
 from globus_cli.services.transfer import TRANSFER_RESOURCE_SERVER, get_client
+from globus_cli.termio import FORMAT_TEXT_RAW, formatted_print
 
 
 class SubscriptionIdType(click.ParamType):
