+++ src/globus_cli/commands/endpoint/show.py
@@ -1,7 +1,7 @@
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command, endpoint_id_arg
-from globus_cli.safeio import FORMAT_TEXT_RECORD, FormatField, formatted_print
 from globus_cli.services.transfer import TRANSFER_RESOURCE_SERVER, get_client
+from globus_cli.termio import FORMAT_TEXT_RECORD, FormatField, formatted_print
 
 
 @command("show")
