+++ src/globus_cli/commands/endpoint/update.py
@@ -5,12 +5,12 @@ from globus_cli.parsing import (
     endpoint_id_arg,
     validate_endpoint_create_and_update_params,
 )
-from globus_cli.safeio import FORMAT_TEXT_RAW, formatted_print
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     assemble_generic_doc,
     get_client,
 )
+from globus_cli.termio import FORMAT_TEXT_RAW, formatted_print
 
 
 @command("update")
