+++ src/globus_cli/commands/session/show.py
@@ -4,8 +4,8 @@ import globus_sdk
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command
-from globus_cli.safeio import formatted_print, print_command_hint
 from globus_cli.services.auth import AUTH_RESOURCE_SERVER, get_auth_client
+from globus_cli.termio import formatted_print, print_command_hint
 from globus_cli.tokenstore import internal_auth_client, token_storage_adapter
