+++ src/globus_cli/commands/task/list.py
@@ -3,12 +3,12 @@ import click
 from globus_cli.login_manager import requires_login
 from globus_cli.paging_wrapper import PagingWrapper
 from globus_cli.parsing import command
-from globus_cli.safeio import formatted_print
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     get_client,
     iterable_response_to_dict,
 )
+from globus_cli.termio import formatted_print
 
 
 def _format_date_callback(ctx, param, value):
