+++ src/globus_cli/commands/task/pause_info.py
@@ -2,8 +2,8 @@ import click
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command, task_id_arg
-from globus_cli.safeio import FORMAT_TEXT_RECORD, formatted_print
 from globus_cli.services.transfer import TRANSFER_RESOURCE_SERVER, get_client
+from globus_cli.termio import FORMAT_TEXT_RECORD, formatted_print
 
 EXPLICIT_PAUSE_MSG_FIELDS = [
     ("Source Endpoint", "source_pause_message"),
