+++ src/globus_cli/commands/task/show.py
@@ -2,12 +2,12 @@ import click
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command, mutex_option_group, task_id_arg
-from globus_cli.safeio import FORMAT_TEXT_RECORD, formatted_print
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     get_client,
     iterable_response_to_dict,
 )
+from globus_cli.termio import FORMAT_TEXT_RECORD, formatted_print
 
 COMMON_FIELDS = [
     ("Label", "label"),
