+++ src/globus_cli/commands/task/update.py
@@ -2,12 +2,12 @@ import click
 
 from globus_cli.login_manager import requires_login
 from globus_cli.parsing import command, task_id_arg
-from globus_cli.safeio import formatted_print
 from globus_cli.services.transfer import (
     TRANSFER_RESOURCE_SERVER,
     assemble_generic_doc,
     get_client,
 )
+from globus_cli.termio import formatted_print
 
 
 @command(
