+++ src/globus_cli/termio/__init__.py
@@ -1,18 +1,18 @@
-from globus_cli.safeio.check_pty import (
+import click
+
+from .context import (
     err_is_terminal,
-    out_is_terminal,
-    term_is_interactive,
-)
-from globus_cli.safeio.errors import PrintableErrorField, write_error_info
-from globus_cli.safeio.get_option_vals import (
     get_jmespath_expression,
     is_verbose,
+    out_is_terminal,
     outformat_is_json,
     outformat_is_text,
     outformat_is_unix,
+    term_is_interactive,
     verbosity,
 )
-from globus_cli.safeio.output_formatter import (
+from .errors import PrintableErrorField, write_error_info
+from .output_formatter import (
     FORMAT_JSON,
     FORMAT_SILENT,
     FORMAT_TEXT_RAW,
@@ -21,7 +21,16 @@ from globus_cli.safeio.output_formatter import (
     FormatField,
     formatted_print,
 )
-from globus_cli.safeio.write import print_command_hint
+
+
+def print_command_hint(message):
+    """
+    Wrapper around echo that checks terminal state
+    before printing a given command hint message
+    """
+    if term_is_interactive() and err_is_terminal() and out_is_terminal():
+        click.echo(click.style(message, fg="yellow"), err=True)
+
 
 __all__ = [
     "print_command_hint",
